crewai_nl2sql/requirements.txt
crewai==0.30.11
langchain==0.1.20
langchain-openai==0.1.20
langchain-community==0.1.20
sqlalchemy==2.0.31
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.8.2
chromadb==0.5.3
tabulate==0.9.0
colorama==0.4.6
